+++ tests/manual/test.py
@@ -8,6 +8,7 @@ pick_round = 2
 if __name__ == '__main__':
 
     espn = PYESPN('nba')
+    espn.load_season_schedule(season=season, load_play_in=True, load_game_odds=True)
     espn.league.load_regular_season_schedule(season=season, load_game_odds=True)
     #espn.load_season_rosters(season=season)
     #espn.teams[0].load_season_roster_box_score(season=season)