+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.4
 * adding in preseason/postseason schedules
+  * also play in for nba
 * 
 
 ## 0.3.3